+++ server/db/update_scripts/22_update_v6.7.0_v6.8.0.sql
@@ -30,9 +30,14 @@ Use the following syntax to achieve this:
 Please use relative paths unless the `url` is an external link.'
 WHERE id_fields = get_field_id('redirect_at_end');
 
+DELETE FROM `styles_fields`
+WHERE id_styles = get_style_id('input') AND id_fields = get_field_id('toggle_switch');
+
 -- Add new style checkbox
 INSERT IGNORE INTO `styles` (`name`, `id_type`, id_group, `description`) VALUES ('checkbox', '1', (select id from styleGroup where `name` = 'Mobile' limit 1), 'Exacute shortcut commands in the mobile app that can open native windows. The functinality is based on [capacitor-native-settings](https://github.com/RaphaelWoude/capacitor-native-settings)');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) VALUES (get_style_id('checkbox'), get_field_id('css'), NULL, 'Allows to assign CSS classes to the root item of the style.');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) VALUES (get_style_id('checkbox'), get_field_id('css_mobile'), NULL, 'Allows to assign CSS classes to the root item of the style for the mobile version.');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) VALUES (get_style_id('checkbox'), get_field_id('condition'), NULL, 'The field `condition` allows to specify a condition. Note that the field `condition` is of type `json` and requires\n1. valid json syntax (see https://www.json.org/)\n2. a valid condition structure (see https://github.com/jwadhams/json-logic-php/)\n\nOnly if a condition resolves to true the sections added to the field `children` will be rendered.\n\nIn order to refer to a form-field use the syntax `"@__form_name__#__from_field_name__"` (the quotes are necessary to make it valid json syntax) where `__form_name__` is the value of the field `name` of the style `formUserInput` and `__form_field_name__` is the value of the field `name` of any form-field style.');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`) VALUES (get_style_id('checkbox'), get_field_id('data_config'), '', 'Define data configuration for fields that are loaded from DB and can be used inside the style with their param names. The name of the field can be used between {{param_name}} to load the required value');
+-- add `toggle_switch` field to style `input`, when enabled and the type is checkbox, then the input will be loaded as toggle switch
+INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('checkbox'), get_field_id('toggle_switch'), 0, 'If enabled and the `type` of the input is `checkbox`, the input will be presented as a `toggle switch`');